Implement draw word for rectangular board outline.
This was hard! Interpretation vs compilation semantics differ for some words and
figuring out the right way to mix the two was tough. Fortunately, pForth
supports binding stack values to local variables which allowed me to make the
draw closure work.

Right now there is no cleanup of vector objects when they are drawn. I expect
this won't be much of a problem; the objects are around 20 bytes each, so it
will take quite a lot of them to make much of a dent in memory. However, it
would be cleaner to remove them somehow when they are no longer needed. Perhaps
an arena/scratchpad could be used to draw them, and reused for subsequent
frames.